+++ test/clj/rems/db/test_user_settings.clj
@@ -2,44 +2,88 @@
   (:require [clojure.test :refer :all]
             [rems.db.testing :refer [rollback-db-fixture test-db-fixture]]
             [rems.db.user-settings :as user-settings]
-            [rems.db.users :as users]))
+            [rems.db.users :as users]
+            [schema.core :as s]))
 
 (use-fixtures :once test-db-fixture)
 (use-fixtures :each rollback-db-fixture)
 
 (deftest test-user-settings
-  (users/add-user! "user1" {:eppn "user1"})
-  (users/add-user! "user2" {:eppn "user2"})
+  (users/add-user! "user" {:eppn "user"})
+  (users/add-user! "unrelated" {:eppn "unrelated"})
 
-  (testing "user settings have the default settings for a new user"
-    (is (= (user-settings/get-user-settings "user1")
-           {:language :en})))
+  (testing "default settings for a new user"
+    (is (= {:language :en
+            :notification-email nil}
+           (s/validate user-settings/UserSettings
+                       (user-settings/get-user-settings "user")))))
 
-  (testing "add a user setting"
-    (user-settings/update-user-settings! "user1" {:language :en})
-    (is (= (user-settings/get-user-settings "user1")
-           {:language :en})))
+  (testing "add settings"
+    (is (= {:success true}
+           (user-settings/update-user-settings! "user" {:language :fi})))
+    (is (= {:language :fi
+            :notification-email nil}
+           (user-settings/get-user-settings "user"))))
 
-  (testing "modify a user setting"
-    (user-settings/update-user-settings! "user1" {:language :fi})
-    (is (= (user-settings/get-user-settings "user1")
-           {:language :fi})))
+  (testing "modify settings"
+    (is (= {:success true}
+           (user-settings/update-user-settings! "user" {:notification-email "user@example.com"})))
+    (is (= {:language :fi
+            :notification-email "user@example.com"}
+           (user-settings/get-user-settings "user"))))
 
-  (testing "updating with empty settings does not change the setting"
-    (user-settings/update-user-settings! "user1" {})
-    (is (= (user-settings/get-user-settings "user1")
-           {:language :fi})))
+  (testing "updating with empty settings does not change settings"
+    (is (= {:success true}
+           (user-settings/update-user-settings! "user" {})))
+    (is (= {:language :fi
+            :notification-email "user@example.com"}
+           (user-settings/get-user-settings "user"))))
 
-  (testing "updating with nil language does not change the setting"
-    (user-settings/update-user-settings! "user1" {:language nil})
-    (is (= (user-settings/get-user-settings "user1")
-           {:language :fi})))
+  (testing "updating with invalid settings does not change settings"
+    (is (= {:success false}
+           (user-settings/update-user-settings! "user" {:notification-email "foo"}))
+        "completely invalid")
+    (is (= {:success false}
+           (user-settings/update-user-settings! "user" {:notification-email "bar@example.com" ; valid
+                                                        :language :de})) ; invalid
+        "partially invalid")
+    (is (= {:language :fi
+            :notification-email "user@example.com"}
+           (user-settings/get-user-settings "user"))))
 
-  (testing "updating with undefined language does not change the setting"
-    (user-settings/update-user-settings! "user1" {:language :de})
-    (is (= (user-settings/get-user-settings "user1")
-           {:language :fi})))
+  (testing "changing one user's settings does not change another user's settings"
+    (is (= {:success true}
+           (user-settings/update-user-settings! "unrelated" {:notification-email "unrelated@example.com"})))
+    (is (= {:success true}
+           (user-settings/update-user-settings! "user" {:notification-email "changed@example.com"})))
+    (is (= {:language :en
+            :notification-email "unrelated@example.com"}
+           (user-settings/get-user-settings "unrelated")))))
 
-  (testing "modifying a setting does not change setting for an unrelated user"
-    (is (= (user-settings/get-user-settings "user2")
-           {:language :en}))))
+(deftest test-language
+  (testing "valid language"
+    (is (= {:language :en} (user-settings/validate-new-settings {:language :en})))
+    (is (= {:language :fi} (user-settings/validate-new-settings {:language :fi}))))
+
+  (testing "undefined language"
+    (is (= {} (user-settings/validate-new-settings {:language :de}))))
+
+  (testing "nil language"
+    (is (= {} (user-settings/validate-new-settings {:language nil})))))
+
+(deftest test-notification-email
+  (testing "valid email"
+    (is (= {:notification-email "somebody@example.com"}
+           (user-settings/validate-new-settings {:notification-email "somebody@example.com"}))))
+
+  (testing "invalid email"
+    (is (= {}
+           (user-settings/validate-new-settings {:notification-email "somebody"}))))
+
+  (testing "empty email will clear the setting"
+    (is (= {:notification-email nil}
+           (user-settings/validate-new-settings {:notification-email ""}))))
+
+  (testing "nil email will clear the setting"
+    (is (= {:notification-email nil}
+           (user-settings/validate-new-settings {:notification-email nil})))))